Release 1.0.6
git-svn-id: svn://svn.code.sf.net/p/libkqueue/code/branches/stable@538 fb4e3144-bc1c-4b72-a658-5bcd248dd7f7

+++ src/common/private.h
@@ -164,7 +164,7 @@ struct kqueue {
     RB_ENTRY(kqueue) entries;
 };
 
-struct knote *  knote_lookup(struct filter *, short);
+struct knote *  knote_lookup(struct filter *, uintptr_t);
 struct knote *  knote_lookup_data(struct filter *filt, intptr_t);
 struct knote *  knote_new(void);
 void        knote_free(struct filter *, struct knote *);
